fix: Make query retryPolicy be respected always
For the moment we have incorrect behavior when we define
retryPolicy and error happens executing query.
The problem is that when we get one of the following errors:
  - context.Canceled
  - context.DeadlineExceeded
  - ErrNotFound
We do not do retries. This is bug.
So, fix it by respecting the retryPolicy in any case.

Use case:
testing a Scylla cluster using 'gocql' and making different
disruptions to it we expect some temporary instabilities in
Scylla's responces and configure retry policy for it.
And when we intentionally terminate a Scylla node we expect
some small period of time of errors, but we catch the
'context.Canceled' exception that breaks the retry logic and
makes us get temporary expected errors as final ones
knowing that in several seconds it will be ok.

Fixes: #101